+++ src/app/not-found-page/not-found-page.component.ts
@@ -1,38 +1,17 @@
-import { Component, ViewEncapsulation } from '@angular/core';
+import { Component, inject, ViewEncapsulation } from '@angular/core';
 import { RouterLink } from '@angular/router';
 import { LocalizationModule } from '../localization-module/localization.module';
 import { LanguageService } from '../shared/services/language-service/language.service';
 import locale from './not-found-page.locale.json';
 
 @Component({
-  standalone: true,
   selector: 'page-not-found',
-  template: `
-    <div class="py-10 mx-14">
-      <h1 class="text-4xl font-semibold mb-8 text-center">
-        {{ locale | translate: 'Title' : lang.language }}
-      </h1>
-      <p class="block description text-center font-medium text-xl mb-10">
-        {{ locale | translate: 'Description' : lang.language }}
-      </p>
-      <div class="block text-center">
-        <!--@router-->
-        <a
-          routerLink="/"
-          class="cursor-pointer rounded text-sm font-medium bg-blue-700 text-white py-2 px-4 outline-none"
-          rel="noreferrer noopener"
-        >
-          {{ locale | translate: 'Button' : lang.language }}
-        </a>
-      </div>
-    </div>
-  `,
+  standalone: true,
+  templateUrl: 'not-found-page.component.html',
   encapsulation: ViewEncapsulation.None,
-  // changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [RouterLink, LocalizationModule],
 })
 export class NotFoundPageComponent {
   protected readonly locale = locale;
-
-  constructor(protected readonly lang: LanguageService) {}
+  protected readonly lang = inject(LanguageService);
 }