+++ CHANGELOG.md
@@ -10,6 +10,18 @@ Run command for (without commit):
 
 ## Versions
 
+### 1.16.3
+
+- refactor `NotFoundPageComponent`
+
+### 1.16.2
+
+- change `HomepageModule` into standalone `HomepageComponent`
+
+### 1.16.1
+
+- add component class into Taiwind
+
 ### 1.16.0
 
 - remove `LanguageSwitch` module, use `LocalizationModule` as replacement